+++ ProyectoVenta/Data/Permiso.cs
@@ -5,21 +5,21 @@ namespace ProyectoVenta.Data;
 
 public partial class Permiso
 {
-    public long IdPermisos { get; set; }
+    public int IdPermisos { get; set; }
 
     public string Descripcion { get; set; }
 
-    public long Salidas { get; set; }
+    public int Salidas { get; set; }
 
-    public long Entradas { get; set; }
+    public int Entradas { get; set; }
 
-    public long Productos { get; set; }
+    public int Productos { get; set; }
 
-    public long Clientes { get; set; }
+    public int Clientes { get; set; }
 
-    public long Proveedores { get; set; }
+    public int Proveedores { get; set; }
 
-    public long Inventario { get; set; }
+    public int Inventario { get; set; }
 
-    public long Configuracion { get; set; }
+    public int Configuracion { get; set; }
 }
